allocator/src/configuration.rs: fix: Sanity check added when looking for NITRO_CLI_INSTALL_DIR env variable. It fails to read config file if env variable is not exist.
If env variable not exist then allocator will try to find the config file in / directory. Tried to build everything manually i bumped into this error in amazon linux and ubuntu instances. Also legacy allocator has the same mechanism but it was not failing. it was failing in Rust so i had to add a sanity check.