fix: fetch-blobs: Clear directory when fetch fails
I found a flaw that when an error occurs it wasn't clearing directories since it creates a directory before we try to fetch. Also i realized that i was using NitroCliResult's custom ok_or_exit_with_errno function which exits process if error occurs. I made a mistake by using it in lib.rs now i fixed it to propagate the error.