fix: fetch-blobs: Changed blobs download path
Since nitro-cli don't have sudo permissions i had to change the download directory for binaries. i found one in users directory named .nitro_cli which it would solve our problem. Also i found a flaw when an error occurs i had to delete the directory aswell but there is a case that we don't allow users to use same name for binaries. When that error occurs i had to filter the errors and avoid clearing the directory.